Add table route, refactor functions, and simplify SVG
Refactored `itemTextWidth` to use `Float` for improved precision. Introduced a new `/table.html` endpoint with corresponding Freemarker template. Simplified SVG content structure and reduced redundant definitions to streamline rendering.

+++ src/main/kotlin/io/docops/docopsextensionssupport/adr/ADRParser.kt
@@ -125,7 +125,7 @@ class ADRParser {
             list.addAll(it.addLinebreaks(config.lineSize))
             list.add("")
         }
-        val itemArray = itemTextWidth(parts.joinToString(), SplitLength, 11)
+        val itemArray = itemTextWidth(parts.joinToString(), SplitLength.toFloat(), 11)
         itemArray.add("")
         return itemArray
     }
@@ -149,7 +149,7 @@ class ADRParser {
             list.addAll(it.addLinebreaks(config.lineSize))
             list.add("")
         }
-        val itemArray = itemTextWidth(parts.joinToString(), SplitLength, 11)
+        val itemArray = itemTextWidth(parts.joinToString(), SplitLength.toFloat(), 11)
         itemArray.add("")
         return itemArray
     }
@@ -173,7 +173,7 @@ class ADRParser {
             list.addAll(it.addLinebreaks(config.lineSize))
             list.add("")
         }
-        val itemArray = itemTextWidth(parts.joinToString(), SplitLength, 11)
+        val itemArray = itemTextWidth(parts.joinToString(), SplitLength.toFloat(), 11)
         itemArray.add("")
         return itemArray
     }
